+++ src/Ops.Controllers/Areas/SiteManagement/ViewModels/Location/LocationPartialViewModel.cs
@@ -1,0 +1,8 @@
+﻿namespace Ocuda.Ops.Controllers.Areas.SiteManagement.ViewModels.Location
+{
+    public class LocationPartialViewModel
+    {
+        public string LocationName { get; set; }
+        public string LocationStub { get; set; }
+    }
+}