+++ src/Ops.Service/Interfaces/Promenade/Repositories/ILocationHoursRepository.cs
@@ -10,6 +10,5 @@ namespace Ocuda.Ops.Service.Interfaces.Promenade.Repositories
         Task<LocationHours> GetByIdsAsync(DayOfWeek dayOfWeek, int locationId);
         Task<List<LocationHours>> GetLocationHoursByLocationId(int locationId);
         Task<bool> IsDuplicateDayAsync(LocationHours locationHour);
-        Task<ICollection<LocationHours>> GetWeeklyHoursAsync(int locationId);
     }
 }